Use modern terminology: rename slaves to agents
Mesos started using 'agent' rather than 'slave' as of Mesos version 1.0:
https://issues.apache.org/jira/browse/MESOS-1478

Regardless of whether Scrappy supports that version, it's 2018 and we
should use modern and appropriate terminology that doesn't alienate
people.

When the old 'slaves' command is used, an error will be printed and the
program will exit. I considered adding a deprecation warning, but I
don't think it's worth it as Scrappy has no official releases and users
should not be surprised by backwards-incompatible changes.