Add awareness of MULTI_ROLE, and hacks to work around mesos behaviours for it.
In particular, if a marathon framework has a default role of 'storage', that framework
still implicitly gets '*' role subscriptions.  Tasks in that marathon can
explicitly uses '*' via setting `acceptedResourceRoles=['*']`, and the task's
'role' in the mesos /master/state will still be 'storage'.

This is a bug in mesos in my view, but it occurs- and was causing scrappy
to segfault when it tried to access a non-existent role for the given slave.

The 'hack' for this is that if the role specified by the task doesn't exist on
the host, assume it's an implicit '*' and use that.